+++ lib/recipes/web-generic/recipe.rb
@@ -12,6 +12,7 @@ WEB_RECIPE = Kitchen::BookRecipe.new(book_short_name: :web) do |doc, resources|
 
   # Web directions
   BakeImages.v1(book_pages: book_pages, resources: resources)
+  BakeOrderHeaders.v1(within: book_pages)
 
   # Restore
   WebPostBakeRestore.v1(book_pages: book_pages)