+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v2.9.0] - 2024-01-12
+
 * Bake `eoc` in `information-systems`
 * Bake `notes` in `information-systems`
 * Bake `footnotes`, `rex wrappers`, and `target labels` in `information-systems`